[Benchmark]add mask2former and vit_adapter to train benchmark (#3111)
* fix bisenetv2 and fastscnn config for benchmark

* [Benchmark] Align the benchmark test environment of each model, such as num_workers, log_iters, epoch, not_eval, to_static_train, etc.

* [Benchmark] add pphumanseg_server to train benchmark

* [Benchmark] add maskformer to train benchmark

* [Benchmark]add mask2former and vit_adapter to train benchmark

* Fix some code

* Fix some bugs and comments

* Fix some bugs and comments

---------

+++ contrib/PanopticSeg/paddlepanseg/models/mask2former/transformer_decoder.py
@@ -336,7 +336,8 @@ class MultiScaleMaskedTransformerDecoder(nn.Layer):
                                  attn_mask_target_size):
         decoder_output = self.decoder_norm(output)
         outputs_class = self.class_embed(decoder_output)
-        mask_embed = self.mask_embed(decoder_output)
+        mask_embed = self.mask_embed(decoder_output).astype('float32')
+        mask_features = mask_features.astype('float32')
         outputs_mask = paddle.einsum('bqc,bchw->bqhw', mask_embed,
                                      mask_features)
